Test resolving generic module parameter.
