+++ server/lsp/parser/parsed_modules_test.go
@@ -3,7 +3,9 @@ package parser
 import (
 	"testing"
 
+	"github.com/pherrymason/c3-lsp/lsp/document"
 	"github.com/pherrymason/c3-lsp/lsp/symbols"
+	idx "github.com/pherrymason/c3-lsp/lsp/symbols"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -98,4 +100,38 @@ func TestParserModules_HasImplicitLoadableModules_should_return_true_when_there_
 	pm.modules["foo"] = &loadableModule
 
 	assert.True(t, pm.HasImplicitLoadableModules(module))
+}
+
+func TestExtractSymbols_module_with_generics(t *testing.T) {
+
+	source := `module foo_test(<Type1, Type2>);
+		struct Foo
+		{
+			Type1 a;
+		}
+		fn Type2 test(Type2 b, Foo *foo)
+		{
+			return foo.a + b;
+		}`
+
+	doc := document.NewDocument("docid", source)
+	parser := createParser()
+	symbols := parser.ParseSymbols(&doc)
+
+	module := symbols.Get("foo_test")
+	assert.Equal(t, "foo_test", module.GetName())
+
+	// Generic parameter was found
+	generic, ok := module.GenericParameters["Type1"]
+	assert.True(t, ok)
+	assert.Equal(t, "Type1", generic.GetName())
+	assert.Equal(t, idx.NewRange(0, 17, 0, 22), generic.GetIdRange())
+	assert.Equal(t, idx.NewRange(0, 17, 0, 22), generic.GetDocumentRange())
+
+	// Generic parameter was found
+	generic, ok = module.GenericParameters["Type2"]
+	assert.True(t, ok)
+	assert.Equal(t, "Type2", generic.GetName())
+	assert.Equal(t, idx.NewRange(0, 24, 0, 29), generic.GetIdRange())
+	assert.Equal(t, idx.NewRange(0, 24, 0, 29), generic.GetDocumentRange())
 }
