+++ server/lsp/symbols/generic_parameter.go
@@ -18,3 +18,7 @@ func NewGenericParameter(name string, module string, docId string, idRange Range
 		),
 	}
 }
+
+func (g GenericParameter) GetHoverInfo() string {
+	return g.GetName()
+}
